Kernel/USB: Don't truncate the max packet size in the MassStorage driver

+++ Kernel/Bus/USB/Drivers/MassStorage/MassStorageDriver.cpp
@@ -102,9 +102,9 @@ ErrorOr<void> MassStorageDriver::initialise_bulk_only_device(USB::Device& device
         dmesgln("SCSI/BBB: WARNING: USB Mass Storage Device supports multiple LUNs ({}) only targetting first LUN", max_luns);
 
     u8 in_pipe_address = 0xff;
-    u8 in_max_packet_size;
+    u16 in_max_packet_size;
     u8 out_pipe_address = 0xff;
-    u8 out_max_packet_size;
+    u16 out_max_packet_size;
 
     if (interface.descriptor().number_of_endpoints < 2) {
         dmesgln("SCSI/BBB: Interface does not provide enough endpoints for advertised Bulk-only transfer protocol; Rejecting");
